moved unit tests into its own classes Fixed unit tests and code for non US setting
