+++ .gitignore
@@ -1,0 +1,6 @@
+﻿################################################################################
+# This .gitignore file was automatically created by Microsoft(R) Visual Studio.
+################################################################################
+
+/DevOpsOverview/Repository/src/.vs/MercuryHealth/v15/.suo
+/DevOpsOverview/Repository/src/.vs
