+++ DevOpsOverview/Repository/src/MercuryHealth.UnitTests/UserRepositoryTests.cs
@@ -8,9 +8,10 @@ using System.Fakes;
 namespace MercuryHealth.UnitTests
 {
     [TestClass]
-    public class HealthCalculatorTests
+    public class UserRepositoryTests
     {
         [TestMethod]
+        [TestCategory("Unit Tests")]
         public void GenerateReportNameTest()
         {
             using (ShimsContext.Create())
@@ -27,7 +28,9 @@ namespace MercuryHealth.UnitTests
                 var calc = new HealthCalculator();
                 var reportName = calc.GenerateReportName(25);
 
-                Assert.AreEqual("2/5/2016 12:00:00 AM: Abel Wang", reportName, "Report name is wrong");
+                var expectedDateTime = new DateTime(2016, 2, 5).ToString();
+
+                Assert.AreEqual(expectedDateTime + ": Abel Wang", reportName, "Report name is wrong");
             }
         }
     }
